requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
Werkzeug==1.0.1
